pyproject.toml
[project]
name = "transport-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.25.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
